# cmd/docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GoFundIt is a collaborative platform enabling families and friends to plan and fund group activities, such as vacations and events. Features end-to-end encryption, secure payment handling, and flexible contribution options.",
        "title": "GoFundIt API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://gofundit.com/support",
            "email": "oyeniyibright@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/campaign/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new campaign with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaign"
                ],
                "summary": "Create Campaign",
                "parameters": [
                    {
                        "description": "Campaign Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CampaignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Campaign"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid inputs, please check and try again",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/campaign/{campaignID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a campaign by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaign"
                ],
                "summary": "Get Campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Campaign"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid campaign ID",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing campaign by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaign"
                ],
                "summary": "Update Campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Campaign Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CampaignUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Campaign"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid inputs",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CampaignRequest": {
            "description": "Campaign creation request structure",
            "type": "object",
            "required": [
                "activities",
                "contributors",
                "description",
                "endDate",
                "images",
                "paymentMethod",
                "startDate",
                "title"
            ],
            "properties": {
                "activities": {
                    "description": "@Description CampaignActivities",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Activity"
                    }
                },
                "contributors": {
                    "description": "@Description Campaign contributors",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Contributor"
                    }
                },
                "cryptoToken": {
                    "description": "@Description Crypto token (required if paymentMethod is crypto)\n@example \"ETH\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CryptoToken"
                        }
                    ]
                },
                "description": {
                    "description": "@Description Detailed campaign description\n@example \"This campaign aims to protect and preserve our forest ecosystems...\"",
                    "type": "string",
                    "minLength": 100
                },
                "endDate": {
                    "description": "@Description Campaign end date (must be after start date)\n@example \"2025-02-03T13:51:06Z\"",
                    "type": "string"
                },
                "fiatCurrency": {
                    "description": "@Description Fiat currency code (required if paymentMethod is fiat)\n@example \"USD\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.FiatCurrency"
                        }
                    ]
                },
                "images": {
                    "description": "@Description Campaign images",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CampaignImage"
                    }
                },
                "paymentMethod": {
                    "description": "@Description Payment method (fiat/crypto/manual)\n@example \"fiat\"",
                    "enum": [
                        "fiat",
                        "crypto",
                        "manual"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PaymentMethod"
                        }
                    ]
                },
                "startDate": {
                    "description": "@Description Campaign start date\n@example \"2025-01-03T13:51:06Z\"",
                    "type": "string"
                },
                "title": {
                    "description": "@Description Campaign title\n@example \"Save the Forests\"",
                    "type": "string",
                    "minLength": 4
                }
            }
        },
        "dto.CampaignUpdateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handlers.BadRequestResponse": {
            "description": "Bad request error response structure",
            "type": "object",
            "properties": {
                "errors": {
                    "description": "Validation errors (optional)\n@example [{\"field\":\"field\",\"error\":\"must be a valid data\"}]"
                },
                "message": {
                    "description": "Error message\n@example Invalid input provided",
                    "type": "string"
                },
                "status": {
                    "description": "Status will always be \"Bad Request\"\n@example Bad Request",
                    "type": "string",
                    "enum": [
                        "Bad Request"
                    ],
                    "example": "Bad Request"
                }
            }
        },
        "handlers.SuccessResponse": {
            "description": "Successful API response structure",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data (optional)\n@example null"
                },
                "message": {
                    "description": "Response message describing the success\n@example Operation completed successfully",
                    "type": "string"
                },
                "status": {
                    "description": "Status will always be \"OK\"\n@example OK",
                    "type": "string",
                    "enum": [
                        "OK"
                    ],
                    "example": "OK"
                }
            }
        },
        "handlers.UnauthorizedResponse": {
            "description": "Unauthorized error response structure",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Will always be \"Unauthorized\"\n@example Unauthorized",
                    "type": "string",
                    "enum": [
                        "Unauthorized"
                    ],
                    "example": "Unauthorized"
                },
                "status": {
                    "description": "Status will always be false for unauthorized\n@example false",
                    "type": "string",
                    "enum": [
                        "Unauthorized"
                    ],
                    "example": "Unauthorized"
                }
            }
        },
        "handlers.ValidationError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "must be a valid field"
                },
                "field": {
                    "type": "string",
                    "example": "field"
                }
            }
        },
        "handlers.response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data (optional)"
                },
                "errors": {
                    "description": "Error details (optional)"
                },
                "message": {
                    "description": "Response message\n@example Operation completed successfully",
                    "type": "string"
                },
                "status": {
                    "description": "Response status\n@example OK",
                    "type": "string"
                }
            }
        },
        "models.Activity": {
            "type": "object",
            "required": [
                "campaignId",
                "cost",
                "title"
            ],
            "properties": {
                "campaignId": {
                    "type": "string"
                },
                "contributors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Contributor"
                    }
                },
                "cost": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isApproved": {
                    "type": "boolean"
                },
                "isMandatory": {
                    "type": "boolean"
                },
                "subtitle": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "minLength": 4
                }
            }
        },
...
        "models.Campaign": {
            "type": "object",
            "required": [
                "activities",
                "contributors",
                "createdByHandle",
                "description",
                "endDate",
                "images",
                "paymentMethod",
                "startDate",
                "targetAmount",
                "title"
            ],
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Activity"
                    }
                },
                "contributors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Contributor"
                    }
                },
                "createdByHandle": {
                    "type": "string"
                },
                "cryptoToken": {
                    "$ref": "#/definitions/models.CryptoToken"
                },
                "description": {
                    "type": "string",
                    "minLength": 100
                },
                "endDate": {
                    "type": "string"
                },
                "fiatCurrency": {
                    "$ref": "#/definitions/models.FiatCurrency"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CampaignImage"
                    }
                },
                "key": {
                    "type": "string"
                },
                "paymentMethod": {
                    "description": "Payment",
                    "enum": [
                        "fiat",
                        "crypto",
                        "manual"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PaymentMethod"
                        }
                    ]
                },
                "payout": {
                    "description": "Payout",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Payout"
                        }
                    ]
                },
                "startDate": {
                    "type": "string"
                },
                "targetAmount": {
                    "type": "number"
                },
                "title": {
                    "type": "string",
                    "minLength": 4
                }
            }
        },
        "models.CampaignImage": {
            "type": "object",
            "required": [
                "imageUrl"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                }
            }
        },
...
        "models.Contributor": {
            "type": "object",
            "required": [
                "amount",
                "campaignId"
            ],
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Activity"
                    }
                },
                "amount": {
                    "type": "number",
                    "minimum": 0
                },
                "campaignId": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                },
                "payment": {
                    "$ref": "#/definitions/models.Payment"
                }
            }
        },
        "models.CryptoAccount": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cryptoToken": {
                    "$ref": "#/definitions/models.CryptoToken"
                }
            }
        },
        "models.CryptoToken": {
            "type": "string",
            "enum": [
                "USDT",
                "USDC",
                "BUSD",
                "DAI"
            ],
            "x-enum-varnames": [
                "USDT",
                "USDC",
                "BUSD",
                "DAI"
            ]
        },
        "models.FiatAccount": {
            "type": "object",
            "properties": {
                "accountName": {
                    "type": "string"
                },
                "accountNumber": {
                    "type": "string"
                },
                "bankCode": {
                    "type": "string"
                },
                "bankName": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
        "models.FiatCurrency": {
            "type": "string",
            "enum": [
                "GHS",
                "NGN"
            ],
            "x-enum-varnames": [
                "GHS",
                "NGN"
            ]
        },
        "models.ManualPaymentProof": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "authorization_url": {
                    "description": "PaymentURL",
                    "type": "string"
                },
                "campaignId": {
                    "type": "string"
                },
                "contributorId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "gatewayResponse": {
                    "type": "string"
                },
                "paymentMethod": {
                    "$ref": "#/definitions/models.PaymentMethod"
                },
                "paymentProof": {
                    "$ref": "#/definitions/models.ManualPaymentProof"
                },
                "paymentStatus": {
                    "$ref": "#/definitions/models.PaymentStatus"
                },
                "reference": {
                    "type": "string"
                }
            }
        },
        "models.PaymentMethod": {
            "type": "string",
            "enum": [
                "fiat",
                "crypto",
                "manual"
            ],
            "x-enum-varnames": [
                "PaymentMethodFiat",
                "PaymentMethodCrypto",
                "PaymentMethodManual"
            ]
        },
        "models.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "succeeded",
                "failed",
                "pending_approval"
            ],
            "x-enum-varnames": [
                "PaymentStatusPending",
                "PaymentStatusSucceeded",
                "PaymentStatusFailed",
                "PaymentStatusPendingApproval"
            ]
        },
        "models.Payout": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "campaignId": {
                    "type": "string"
                },
                "completedAt": {
                    "type": "string"
                },
                "cryptoAccount": {
                    "$ref": "#/definitions/models.CryptoAccount"
                },
                "failureReason": {
                    "type": "string"
                },
                "fiatAccount": {
                    "$ref": "#/definitions/models.FiatAccount"
                },
                "payoutMethod": {
                    "$ref": "#/definitions/models.PaymentMethod"
                },
                "processedAt": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.PayoutStatus"
                }
            }
        },
        "models.PayoutStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "PayoutStatusPending",
                "PayoutStatusProcessing",
                "PayoutStatusCompleted",
                "PayoutStatusFailed"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for accessing the API endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token for authenticated requests. Use the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CampaignKeyAuth": {
            "description": "Campaign Key required for campaign-specific operations",
            "type": "apiKey",
            "name": "Campaign-Key",
            "in": "header"
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
...
}